requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
Pillow==10.4.0
openai==1.52.0
